requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27
tenacity>=8.0
streamlit>=1.35
anthropic>=0.30
openai>=1.0
python-dotenv>=1.0
pyyaml>=6.0
loguru>=0.7
yfinance>=0.2
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.0
respx>=0.21
